+++ drasyl-core/src/test/java/org/drasyl/pipeline/SimpleDuplexHandlerTest.java
@@ -28,6 +28,7 @@ import org.drasyl.identity.CompressedPublicKey;
 import org.drasyl.identity.Identity;
 import org.drasyl.peer.connection.message.ApplicationMessage;
 import org.drasyl.peer.connection.message.ChunkedMessage;
+import org.drasyl.pipeline.codec.ApplicationMessage2ObjectHolderHandler;
 import org.drasyl.pipeline.codec.DefaultCodec;
 import org.drasyl.pipeline.codec.ObjectHolder;
 import org.drasyl.pipeline.codec.ObjectHolder2ApplicationMessageHandler;
@@ -99,6 +100,7 @@ class SimpleDuplexHandlerTest {
                     identity,
                     TypeValidator.ofInboundValidator(config),
                     TypeValidator.ofOutboundValidator(config),
+                    ApplicationMessage2ObjectHolderHandler.INSTANCE,
                     DefaultCodec.INSTANCE, handler);
             final TestObserver<Pair<CompressedPublicKey, Object>> inboundMessageTestObserver = pipeline.inboundMessages().test();
             final TestObserver<ApplicationMessage> outboundMessageTestObserver = pipeline.outboundMessages(ApplicationMessage.class).test();
@@ -141,7 +143,9 @@ class SimpleDuplexHandlerTest {
                     identity,
                     TypeValidator.ofInboundValidator(config),
                     TypeValidator.ofOutboundValidator(config),
-                    ObjectHolder2ApplicationMessageHandler.INSTANCE, DefaultCodec.INSTANCE, handler);
+                    ApplicationMessage2ObjectHolderHandler.INSTANCE,
+                    ObjectHolder2ApplicationMessageHandler.INSTANCE,
+                    DefaultCodec.INSTANCE, handler);
             final TestObserver<Pair<CompressedPublicKey, Object>> inboundMessageTestObserver = pipeline.inboundMessages().test();
             final TestObserver<ApplicationMessage> outboundMessageTestObserver = pipeline.outboundMessages(ApplicationMessage.class).test();
 
@@ -191,7 +195,9 @@ class SimpleDuplexHandlerTest {
                     identity,
                     TypeValidator.ofInboundValidator(config),
                     TypeValidator.ofOutboundValidator(config),
-                    ObjectHolder2ApplicationMessageHandler.INSTANCE, DefaultCodec.INSTANCE, handler);
+                    ApplicationMessage2ObjectHolderHandler.INSTANCE,
+                    ObjectHolder2ApplicationMessageHandler.INSTANCE,
+                    DefaultCodec.INSTANCE, handler);
             final TestObserver<Pair<CompressedPublicKey, Object>> inboundMessageTestObserver = pipeline.inboundMessages().test();
             final TestObserver<ApplicationMessage> outboundMessageTestObserver = pipeline.outboundMessages(ApplicationMessage.class).test();
             final TestObserver<Event> eventTestObserver = pipeline.inboundEvents().test();
@@ -239,7 +245,9 @@ class SimpleDuplexHandlerTest {
                     identity,
                     TypeValidator.ofInboundValidator(config),
                     TypeValidator.ofOutboundValidator(config),
-                    ObjectHolder2ApplicationMessageHandler.INSTANCE, DefaultCodec.INSTANCE, handler);
+                    ApplicationMessage2ObjectHolderHandler.INSTANCE,
+                    ObjectHolder2ApplicationMessageHandler.INSTANCE,
+                    DefaultCodec.INSTANCE, handler);
             final TestObserver<Pair<CompressedPublicKey, Object>> inboundMessageTestObserver = pipeline.inboundMessages().test();
             final TestObserver<ApplicationMessage> outboundMessageTestObserver = pipeline.outboundMessages(ApplicationMessage.class).test();
             final TestObserver<Event> eventTestObserver = pipeline.inboundEvents().test();